requirements.txt
streamlit
pdfplumber
langchain
langchain-community
chromadb
sentence-transformers
huggingface-hub
python-dotenv
langchain-text-splitters